requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
pytest
pytest-flask
python-dotenv
